Filters Refactor, FilterSection implementation
