+++ src/pages/BookingsPage/components/BookingsTable/styles/BookingsTable.module.css
@@ -1,0 +1,4 @@
+.guestRowHeader:hover{
+    cursor: pointer;
+    background-color: rgb(200, 200, 200);
+}
